refactor(products): Name list price types by convention
Changes:
- Rename the feature-only product list price type module to use the types suffix
- Update every product list price type import

Keep the new feature types aligned with the repository's colocated type naming convention.

Notes:
- Review maintainability finding for product-list-price-types.ts (reviews/REVIEW-2026-08-02-PRODUCT-LOCAL-PRICE-FILTER-BY-AREA.md)
