+++ frontend/src/app/products/typings/product-list-price-types.ts
